+++ src/components/Input/BigInput.tsx
@@ -1,6 +1,6 @@
 import styled, { css } from "styled-components";
 
-import getColor from "../../styles/themeHelpers/getColor";
+import getColorByLocation from "../../styles/themeHelpers/getColorByLocation";
 import getFontFamily from "../../styles/themeHelpers/getFontFamily";
 import { getBreakpoint } from "../../styles/utils/responsive";
 import { margin, MarginProps } from "../../styles/utils/spacing";
@@ -9,7 +9,9 @@ import UnstyledInput, { UnstyledInputProps } from "../UnstyledInput";
 const inputStyles = css`
   height: ${(props) => props.theme.bigInput.height};
   border-radius: ${(props) => props.theme.bigInput.borderRadius};
-  border-color: ${getColor((props) => props.palette.bigInput.default.border)};
+  border-color: ${getColorByLocation(
+    ({ theme }) => theme.bigInput.states.default.border
+  )};
   border-width: ${(props) => props.theme.bigInput.borderWidth};
   border-style: solid;
   padding-left: 12px;
@@ -30,11 +32,15 @@ const inputStyles = css`
   }
 
   :valid:not([value=""]) {
-    border-color: ${getColor((props) => props.palette.bigInput.valid.border)};
+    border-color: ${getColorByLocation(
+      ({ theme }) => theme.bigInput.states.valid.border
+    )};
   }
 
   ::placeholder {
-    color: ${getColor((props) => props.palette.bigInput.valid.placeholder)};
+    color: ${getColorByLocation(
+      ({ theme }) => theme.bigInput.states.valid.placeholder
+    )};
     opacity: 1;
   }