+++ src/components/SiteHeader/SiteHeader.test.tsx
@@ -1,27 +1,11 @@
-import { screen, fireEvent } from "@testing-library/react";
-
-import { UserStyleContextProvider } from "../../context/UserStyleContext";
 import renderWithProviders from "../../__tests__/__helpers__/renderWithProviders";
 
-import SiteHeader from "./SiteHeader";
+import SiteHeader from ".";
 
-function renderUserContext() {
-  return renderWithProviders(
-    <UserStyleContextProvider>
-      <SiteHeader />
-    </UserStyleContextProvider>
-  );
-}
+describe("components/SiteHeader", () => {
+  test("header should be in the document", () => {
+    const { getByRole } = renderWithProviders(<SiteHeader />);
 
-test("Renders pupil as default user", () => {
-  renderUserContext();
-  expect(screen.getByText("pupils")).toBeInTheDocument();
-});
-
-test("User changes to teacher when button is clicked", () => {
-  renderUserContext();
-  const buttonElement = screen.getByRole("button", { name: /pupils/i });
-  fireEvent.click(buttonElement);
-
-  expect(screen.getByText("teachers")).toBeInTheDocument();
+    expect(getByRole("banner")).toBeInTheDocument();
+  });
 });